Migrate accounts+capabilities to new JMAP authentication spec
According to the recent updates to the JMAP spec, account data is now loaded from the authentication response. Therefore the AuthAccess model was significantly changed to serve as a container for authentication data, account information and server capabilities.

This change requires authentication responses to comply with the latest spec as the model requires certain new properties to be present.

Added: jmap.ServerCapablities
Removed: jmap.AccountCapabilities

The namespace URIs used for the various capabilites are defined in jmap.Constants but not yet final.

see links in issue #58